Gemini Responder Server Code/requirements.txt
Flask
requests
openai-whisper
google-generativeai